+++ src/crud/BooksAuthorsFunc.java
@@ -71,6 +71,14 @@ public class BooksAuthorsFunc extends DBSCon implements BooksAuthorsAccess {
     @Override
     public boolean updateBookAuthor(BooksAuthors BooksAuthors) {
         try (Connection connection = connect()) {
+            int AuthorID = BooksAuthors.getBookAuthor().getAuthorId();
+            int BookID = BooksAuthors.getAuthorBook().getBookID();
+            BooksAuthors existingBooksAuthors = getBookAndAuthorByID(AuthorID, BookID);
+
+            if (existingBooksAuthors == null) {
+                return false;
+            }
+
             try (PreparedStatement stmnt = connection.prepareStatement("UPDATE BooksAuthors SET BookID=? WHERE AuthorID=?")) {
                 stmnt.setInt(1, BooksAuthors.getAuthorBook().getBookID());
                 stmnt.setInt(2, BooksAuthors.getBookAuthor().getAuthorId());
@@ -90,6 +98,12 @@ public class BooksAuthorsFunc extends DBSCon implements BooksAuthorsAccess {
     @Override
     public boolean deleteBookAuthor(int AuthorID, int BookID) {
         try (Connection connection = connect()) {
+            BooksAuthors existingBooksAuthors = getBookAndAuthorByID(AuthorID, BookID);
+
+            if (existingBooksAuthors == null) {
+                return false;
+            }
+
             try (Statement stmnt = connection.createStatement()) {
                 stmnt.execute("DELETE FROM BooksAuthors WHERE AuthorID = '" + AuthorID + "' AND BookID = " + BookID);
             }
